move apt rpc api to single file
